requirements.txt
comet_ml==3.32.4
h5py==3.8.0
hdf5plugin==4.1.1
hydra-core==1.3.2
matplotlib==3.7.1
numpy==1.24.2
omegaconf==2.3.0
pytorch-lightning==2.0.0
scikit-image==0.18.3
pytorch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tqdm==4.64.1
transformers==4.27.2
